+++ src/in-house-bot/types/index.d.ts
@@ -176,6 +176,7 @@ export interface IPBApp extends ITradleObject {
   context: string
   forms?: ApplicationSubmission[]
   verifications?: ApplicationSubmission[]
+  reviewer?: ResourceStub
   relationshipManagers?: ResourceStub[]
   checks?: ResourceStub[]
   status: string